Cambiar los tipos gradoId e idDocente de Integer a Long en las clases DTO

+++ src/main/java/com/sime/backwebsime/DTO/AlumnoCrearDTO.java
@@ -16,5 +16,5 @@ public class AlumnoCrearDTO {
     private String genero;
     private String direccion;
     private String telefonoEmergencia;
-    private Integer gradoId; // ID del grado seleccionado
+    private Long gradoId; // ID del grado seleccionado
 }
